core: migrate stdcm to new infra
Change summary:
* Migrate train simulation code to work with the new infra interfaces
* Migrate STDCM to work with the new infra interfaces
* Run the STDCM pathfinding on blocks instead of routes
* Migrate tests, using a new class to build simplified infras
* Cleanup some legacy code (not all of it)

+++ core/src/main/java/fr/sncf/osrd/utils/graph/functional_interfaces/AStarHeuristic.java
@@ -4,5 +4,5 @@ package fr.sncf.osrd.utils.graph.functional_interfaces;
  * It takes an edge and an offset on this edge as inputs, and returns an estimation of the remaining distance. */
 @FunctionalInterface
 public interface AStarHeuristic<EdgeT> {
-    double apply(EdgeT edge, double offset);
+    double apply(EdgeT edge, long offset);
 }
